refactor: change styles for cart page
