+++ src/components/Cart/CartEmpty.module.scss
@@ -1,0 +1,14 @@
+@import '../../styles/global';
+
+.cartEmpty {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  width: 70vw;
+  height: 40vh;
+  color: $color-brown;
+  font-weight: 700;
+  font-size: 5rem;
+  margin: 0.3rem 1rem;
+}
