--- src/components/Cart/CartForm.module.scss
@@ -1,118 +1,0 @@
-@import '../../styles/global';
-
-.cart_span {
-  color: $color-brown;
-  font-weight: 100;
-  font-size: 1.4rem;
-  margin: 0.1rem 1rem;
-}
-
-.cart_img {
-  width: 107px;
-  height: 107px;
-  margin: 0px 1vw;
-}
-
-.list_cart {
-  margin: 0.3rem;
-  padding: 0.5rem;
-  border-color: $color-brown;
-  border: 0.05vw solid $color-brown;
-  border-radius: 8px;
-}
-
-.list_cart_white {
-  margin: 0.3rem;
-  padding: 0.5rem;
-  background-color: white;
-  border-color: $color-brown;
-  border: 0vw solid $color-brown;
-  border-radius: 8px;
-}
-
-.cart_emtpy {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 70vw;
-  height: 40vh;
-  color: $color-brown;
-  font-weight: 700;
-  font-size: 5rem;
-  margin: 0.3rem 1rem;
-}
-
-.cart_middle {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-  color: #333;
-  font-size: 2vw;
-  padding: 5px;
-  background: transparent;
-}
-
-.cart_between_row {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: space-between;
-  color: #333;
-  font-size: 2vw;
-  padding: 5px;
-  background: transparent;
-}
-
-.cart_small_row {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: start;
-  color: #333;
-  font-size: 2vw;
-  padding: 5px;
-  background: transparent;
-}
-
-.cart_middle_row {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: start;
-  color: #333;
-  font-size: 2vw;
-  padding: 5px;
-  background: transparent;
-}
-
-.cart_ScrollPanel {
-  width: 70%;
-  height: 540px;
-}
-
-@media (max-width: 700px) {
-
-  .cart_middle_row {
-    flex-direction: column-reverse;
-    align-items: center;
-  }
-
-  .cart_ScrollPanel {
-    width: 100%;
-  }
-  
-}
-
-@media (max-width: 500px) {
-
-  .cart_between_row {
-    flex-direction: column;
-  }
-  
-  .cart_span {
-    font-size: 1.1rem;
-  }
-
-}
