+++ src/components/Cart/CartList.tsx
@@ -2,8 +2,7 @@ import { Button } from 'primereact/button';
 import { ScrollPanel } from 'primereact/scrollpanel';
 import { useEffect, useRef, useState } from 'react';
 import { count } from '../../constants/registratForm';
-// import { cartAll, cartDeleteID } from '../../api/customerCart';
-import styles from './CartForm.module.scss';
+import styles from './CartList.module.scss';
 import { LineItem } from '@commercetools/platform-sdk';
 import { useCartID } from './useCart';
 import ItemsVision from './ItemsVision';
@@ -16,6 +15,7 @@ import { Dialog } from 'primereact/dialog';
 
 export const cartData: LineItem[] = [];
 let sumaCart = 0;
+
 export default function CartList(props: { onOffForm: object }): JSX.Element {
   const [visibleCartList] = useState(props.onOffForm);
   const [itemsCart] = useState(cartData);
@@ -73,65 +73,70 @@ export default function CartList(props: { onOffForm: object }): JSX.Element {
   return (
     <div className={styles.list_cart}>
       <div className={styles.cart_middle} style={visibleCartList}>
-        <div className="card">
-          {itemCart.response ? (
-            itemCart.response.length ? (
-              <div className={styles.cart_middle_row}>
-                <ScrollPanel className={styles.cart_ScrollPanel}>
-                  <div className="mb-5">
-                    {itemsCart.map(items => (
-                      <div className={styles.list_cart_white} key={items.id}>
-                        <ItemsVision
-                          value={{
-                            name: items.name['en-US'],
-                            id: items.id,
-                            price: items.price.value.centAmount,
-                            count: items.quantity,
-                            version: itemCart.version,
-                            img: items.variant.images
-                              ? items.variant.images?.[FIRST_INDEX].url
-                              : '',
-                          }}
-                          editDataCart={editData}
-                        />
-                      </div>
-                    ))}
-                  </div>
-                </ScrollPanel>
-                <div className={styles.list_cart_white}>
-                  <p className={styles.cart_span}>Data Cart</p>
-                  <p className={styles.cart_span}>
-                    Suma:&nbsp;
-                    <span className="cart_span" style={{ color: 'red' }}>
-                      {(sumCart / 100).toFixed(2)}
-                    </span>
-                  </p>
-                  <Toast ref={toast} />
-                  <ConfirmPopup
-                    visible={visible}
-                    onHide={(): void => setVisible(false)}
-                    message="Если выбор меньше чем из трёх, это шантаж!)))"
-                    icon="pi pi-exclamation-triangle"
-                    accept={accept}
-                    reject={reject}
-                  />
-                  <div className="card flex justify-content-center">
-                    <Button
-                      ref={buttonEl}
-                      onClick={(): void => setVisible(true)}
-                      icon="pi pi-check"
-                      label="Delete all product"
+        {itemCart.response ? (
+          itemCart.response.length ? (
+            <div className={styles.cart_middle_row}>
+              <div className="mb-5">
+                {itemsCart.map(items => (
+                  <div className={styles.list_cart_white} key={items.id}>
+                    <ItemsVision
+                      value={{
+                        name: items.name['en-US'],
+                        id: items.id,
+                        price: items.price.value.centAmount,
+                        count: items.quantity,
+                        version: itemCart.version,
+                        img: items.variant.images
+                          ? items.variant.images?.[FIRST_INDEX].url
+                          : '',
+                      }}
+                      editDataCart={editData}
                     />
                   </div>
+                ))}
+              </div>
+
+              <div className={styles.list_cart_white}>
+                <p className={styles.cart_span}>Data Cart</p>
+                <p className={styles.cart_span}>
+                  Suma:&nbsp;
+                  <span className="cart_span" style={{ color: 'red' }}>
+                    {(sumCart / 100).toFixed(2)}
+                  </span>
+                </p>
+                <Toast ref={toast} />
+                <ConfirmPopup
+                  visible={visible}
+                  onHide={(): void => setVisible(false)}
+                  message="Если выбор меньше чем из трёх, это шантаж!)))"
+                  icon="pi pi-exclamation-triangle"
+                  accept={accept}
+                  reject={reject}
+                />
+                <div className={styles.button}>
+                  <Button
+                    ref={buttonEl}
+                    onClick={(): void => console.log('12')}
+                    icon="pi pi-check"
+                    label="Proceed to Checkout"
+                  />
+                  <Button
+                    ref={buttonEl}
+                    onClick={(): void => setVisible(true)}
+                    severity="danger"
+                    icon="pi pi-check"
+                    label="Delete all product"
+                  />
                 </div>
               </div>
-            ) : (
-              <CartEmpty />
-            )
+            </div>
           ) : (
             <CartEmpty />
-          )}
-        </div>
+          )
+        ) : (
+          <CartEmpty />
+        )}
+
         <Dialog
           className={styles.module__window}
           style={{ maxWidth: '340px' }}
@@ -144,51 +149,6 @@ export default function CartList(props: { onOffForm: object }): JSX.Element {
           <p>{count.errors}</p>
         </Dialog>
       </div>
-
-      {/* <Button
-        label="Testing Cart"
-        className="mt-3 mb-1"
-        onClick={(): void => {
-          //=====================Запуск запросов для проверок и корректировок=========
-          //==========cartDeleteID
-          // (async (): Promise<void> => {
-          //   await cartDeleteID('bb279157-6fc7-4350-afe5-58be2394d310', 4) // версия в удаляемой корзине
-          //     .then(({ body }) => {
-          //       console.log(body);
-          //       console.log('444444');
-          //     })
-          //     .catch(console.error);
-          // })();
-          //==============cartAll
-          (async (): Promise<void> => {
-            await cartAll()
-              .then(({ body }) => {
-                console.log(body.results);
-                console.log(count.ID);
-                console.log('!!======', count.versionCart, count.cartID);
-                // console.log(
-                //   '!!======',
-                //   body.results[0].id,
-                //   body.results[0].version,
-                // );
-                // for (let i = 4; i < 19; i++) {
-                // (async (): Promise<void> => {
-                //   await cartDeleteID(
-                //     body.results[0].id,
-                //     body.results[0].version,
-                //   ) // версия в удаляемой корзине
-                //     .then(({ body }) => {
-                //       console.log(body);
-                //       console.log('444444');
-                //     })
-                //     .catch(console.error);
-                // })();
-                // }
-              })
-              .catch(console.error);
-          })();
-        }}
-      /> */}
     </div>
   );
 }
