+++ src/components/Cart/ItemsVision.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { IpropsItems } from '../../types/interface';
-import styles from './CartForm.module.scss';
+import styles from './ItemsVision.module.scss';
 import {
   InputNumber,
   InputNumberValueChangeEvent,
@@ -31,34 +31,28 @@ export default function ItemsVision(props: IpropsItems): JSX.Element {
       <div>
         <div className={styles.cart_small_row}>
           <img className={styles.cart_img} src={props.value.img} alt="" />
-          <div className="card flex justify-content-center">
-            <InputNumber
-              value={value}
-              onValueChange={(e: InputNumberValueChangeEvent): void => {
-                if (e.value) {
-                  setValue(e.value);
-                  asyncUpdateItemCart(props.value.id, e.value, callback);
-                }
-              }}
-              showButtons
-              buttonLayout="vertical"
-              style={{ width: '3rem' }}
-              incrementButtonIcon="pi pi-plus"
-              decrementButtonIcon="pi pi-minus"
-            />
-          </div>
-          <div>
-            <p className={styles.cart_span}>
+          <InputNumber
+            className={styles.input}
+            value={value}
+            onValueChange={(e: InputNumberValueChangeEvent): void => {
+              if (e.value) {
+                setValue(e.value);
+                asyncUpdateItemCart(props.value.id, e.value, callback);
+              }
+            }}
+            showButtons
+            buttonLayout="vertical"
+            incrementButtonIcon="pi pi-plus"
+            decrementButtonIcon="pi pi-minus"
+          />
+          <div className={styles.cart_span}>
+            <p>
               Price:&nbsp;
-              <span className="cart_span" style={{ color: 'red' }}>
-                {(props.value.price / 100).toFixed(2)}
-              </span>
+              <span>{(props.value.price / 100).toFixed(2)}</span>
             </p>
-            <p className={styles.cart_span}>
+            <p>
               Total:&nbsp;
-              <span className="cart_span" style={{ color: 'red' }}>
-                {((props.value.price * value) / 100).toFixed(2)}
-              </span>
+              <span>{((props.value.price * value) / 100).toFixed(2)}</span>
             </p>
           </div>
         </div>
@@ -73,6 +67,7 @@ export default function ItemsVision(props: IpropsItems): JSX.Element {
           className="mt-3 mb-1 border-round-lg"
           label="Delete"
           type="submit"
+          severity="danger"
           onClick={(): void => {
             asyncUpdateItemCart(props.value.id, 0, callback);
           }}
