+++ src/pages/Cart/Cart.module.scss
@@ -1,7 +1,11 @@
 @import '../../styles/global';
 
 .page {
-  display: flex;
+  max-width: 1280px;
+  padding: 0 50px;
+  margin: 0 auto;
   min-height: calc(100vh - 150px);
-  padding: 20px;
+  @include media-mobile-big {
+    padding: 0 15px;
+  }
 }
